Split opportunity scorer out of pipelines pairing (burn edge 4)
The census-facing build_uei_pairs / pair_filter_s1 boundary carries a per-file
pipelines label and is imported by the evidence-tier census; it kept the
exploratory topical scorer off that path only via a lazy in-function import.
Move the opportunity-ranking entry points (pair_filter_s2, pair_filter_s3 and
their exclusive helpers _prepare_opportunities, _with_pair_metadata,
_agency_match_level, the notice-type constants, and PAIR_OPPORTUNITY_COLUMNS)
into a new exploratory sibling module opportunity_pairing.py that imports the
scorer at module top level. pairing.py no longer imports .similarity at all.

build_uei_pairs, pair_filter_s1, PAIR_COLUMNS, and every census-touched helper
are unchanged and in place; the new module reuses the shared deterministic
helpers (_normalize, _prepare_priors, _prior_identity) rather than forking them.
Update the exploratory candidate assets, the opportunity script, and the
opportunity test to import from the new module. Remove the tier-allowlist entry.
